Trying to build with current ecosystem
Currently limiting to GHC 7.6.3 and Cabal 1.18, just to get the
dependencies corrected for this to build with the old toolchain...then
we can see about updating to newer GHC and maybe using Stack.

Resolved most dependency conflicts, but have reached a dead-end with
`reader` in `Tabula.Options`. Hoogle's best guess is this is either
part of `mtl` (`Control.Monad.Reader`) or `transformers`
(`Control.Monad.Trans.Reader`): however, neither of these seem to work
and just raise numerous type errors :P